refactor: correction in packages and files of the dependent entity
