+++ src/main/java/br/com/matheusgusmao/incometax/infra/persistence/repository/DependentRepository.java
@@ -1,6 +1,6 @@
 package br.com.matheusgusmao.incometax.infra.persistence.repository;
 
-import br.com.matheusgusmao.incometax.infra.persistence.entity.declaration.DependentEntity;
+import br.com.matheusgusmao.incometax.infra.persistence.entity.dependent.DependentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
